Update Combat log
Update Combat log: now will be uploaded information about  all combat units, not only player. It is damage done, heal done, damage taken and resource recoveries

+++ src/Parser/CombatAnalysis.WoW.CombatParser/Entities/Combat.cs
@@ -6,11 +6,11 @@ public class Combat
 
     public long DamageDone { get; set; }
 
-    public int HealDone { get; set; }
+    public long HealDone { get; set; }
 
-    public int DamageTaken { get; set; }
+    public long DamageTaken { get; set; }
 
-    public int ResourcesRecovery { get; set; }
+    public long ResourcesRecovery { get; set; }
 
     public bool IsWin { get; set; }
